Set up Spell Script and start prefabs
