--- Assets/Scripts/Weapons/Spell.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Spell : MonoBehaviour
-{
-    public int NumUses;
-    public Player Owner;
-    public float duration;
-    public float startTime;
-    
-    void Start()
-    {
-
-    }
-
-    void Update()
-    {
-
-    }
-}
