+++ Assets/Scripts/Weapons/Spell.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 87244686e3bf185428f3a9358fee0c53
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
