+++ Assets/Scripts/Weapons/Spell/InvisibilitySpell.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd29680922399354bbb9831fba9b211b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
