+++ Assets/Scripts/Weapons/Spell/Spell.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Spell : Item
+{
+    public float duration;
+    public float startTime;
+    
+    void Start()
+    {
+
+    }
+
+    void Update()
+    {
+
+    }
+}
