+++ Assets/Scripts/Weapons/Spell/TeleportSpell.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96e103941eff15646beb8009ec38a827
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
